Gallery2: change AES encrypt and decrypt algorithm for baidu yun
When user set AutoUpload to Baidu Yun,Gallery and BaiduNetdisk use the
AES encryption algorithm to encrypt and decrypt the status of AutoUpload.
But the provider for secureRandom has been changed to OpenSSL instead of
Crypto on android 4.2.So the encrypt and decrypt method can't work
normally.And the status which has been encrypted is saved in
BaiduNetdisk's database.And then baiduNetdisk will decrypt and encrpt
the status again.

Change the default provider for secureRandom to AES encryption algorithm
to make it can work on Android 4.2.But also need Baidu to change codes
to adapt current QRD platform.

CRs-Fixed: 502096

Change-Id: Ia1970f00a67678be29ae150ff776af13302d717b