Gallery2: fix IllegalArgumentException when crop small pictures
When user crops a picture, height or width of the crop bounds may be
less than zero or equal to zero if the picture is too small.Then
IllegalArgumentException happens when you drag the grid to crop
picture.

On this situation,we generate original bitmap as the cropped bitmap.

CRs-Fixed: 510658

Change-Id: Ie102fd25933c41c301e91ef64dc3b6c4425a1126