Gallery2: settings of gallery has no options which is blank
There is no code to make the interface of settings be
displayed to users.So we delete settings' item directly.
And PicasaSource.requestSync(activity) will be called
after we click the picasa item. but in
PicasaSource.requestSync, nothing can be done.so it is
unnecessary.

Delete those two items directly.

CRs-Fixed: 351854

Change-Id: Ie05c9b751a15b955a6e2b3d567fde783b66f0e7f