Gallery2: close the option menu when the albumsetpage is paused
If we tap "Menu", and then change language, return back to
gallery, the gallery activity will be recreated, then GLRootView
will be attached to the window, this causes the GLRootView to
be placed above the option menu, so option menu view can not been
seen.

We close the option menu when activity is paused to avoid
this issue.

CRs-Fixed: 510611

Change-Id: I1e76fd20353ee1259a7eddd354659ae9ca3576d6