Camera: Aquire surface texture only when dimensions have changed.
Surface texture creating code is executed in a synchronized fashion
in CameraScreenNail.java file. So UI rendering code will aquire the
lock and try to create the surface. Meanwhile VideoModule receives
orientation change and it calls setParams. During SetParams if
surface object is null getSurfaceTexture call is made for the same
height and width. This control flow also calls mLock.wait. The fix
is made such that the second call to create texture will not issued
 if the older call was also for the same height and width.

Change-Id: I090a91297552f27d9b843a1a77864b92420ae9f3
CRs-Fixed: 518491