Camera: Update skin tone dialog for all scene modes and colour effects.
Skin tone seek bar visibility was getting updated only when scene mode
was set to portrait or party. So when user switch between cameras though
the new camera scene mode is other than the above it will not hide the
seek bar. With the present change it will update the UI always.

Change-Id: I693ce740b3166250841f71d2933021095aa03301
CRs-fixed: 521203