Camera: Gray out unsupported options when scene mode is not set to Auto
When scene mode is set to a value other than Auto, camera application is
querying for the focus, autoExposure, flash, white balance, exposure
values. Framework is returning the values which user had set irrespective
of whether they are supported or not. And also there was no check during
the initalization of the settings dialogs. This change will enable the
above settings only if scene mode is set to auto.

Change-Id: Ieacc94c44fc13a2e9ffce7840771a2e9441b8787
CRs-fixed: 521208