Gallery2: Fix screen freeze when switch language several times
The size of thread pool is 4, when change language, the activity
will be recreate. All of activitystate in state stack also
be recreate. But it only call oncreate, don't call onresume
and onpause. All of threads is submmited at onresume except
TileImageView. TileDecoder of TileImageView will submmited at
oncreate. and it will not be release as it didn't call onpause.
So it will submmit a new thread every time change language.
The thread pool will full after switch several times language.
It will lead other thread can't be submmited, such as slideshow
load thread, it will block the slideshow.

When user touch screen or click back key, it will exit slideshow
it will release the thread, but the thread is not running,
so it will wait the thread complete. So the UI will be block
and the screen is freeze.

Remove TileImageView submit thread at constructor.
Only submit in prepareTextures().

CRs-Fixed: 529641

Change-Id: I5e22f43747460f7f796d96bf6fd3f6f66103adb4