Camera: Fix the countdown timer values updated on UI.
The count down timer values shown on the UI are not reflecting
the values actually selected.
For eg: If user selects 6 seconds, the value selected is 10 seconds.
This is because the code populating the UI is not picking up the
expected duration value. Fix this by showing the correct duration
value on the UI.

Change-Id: I340f811a02f3b90bdda6b071bbce6a2488d3fab8
CRs-Fixed: 524906