CameraApp: Updating zoom value from camera device parameters.
When user switches from one camera to other the zoom values were
stored in a class variable. Due to which the application was using
the same zoom value for different preview size cameras. And which
was making the app to apply same zoom value for two different sizes
while rendering. This change will enable the app to use different
zoom values for different preview sizes.

Change-Id: Ic30001be73e9ddebc503ed1285a74bcd5c0c4780
CRs-fixed: 524903