Gallery2: Remove the share button when preview video with a bad uri
MMS record video the cache file will be specified output path, after
recording back from the video camera activity will looking for the
specified path in "onActivityResult" method. If we share the video
from the movie activty, you can't execute the process in "onActivityResult"
method, so we can't find the video.

In the movie activity, we check the video can be shared and can be
opened through the uri. If not, the share button will be removed.

CRs-Fixed: 538676

Change-Id: I8602a6ac87044a344744390987a299d2f1d06b0b