CameraApp: Updating the view for new list item preference
User has launched the popup menu view in portait mode and can see
all the list items in a single view.

When user turns the device to landscape mode, the popup view will
be shown partially, thus user has to scroll the popup menu to see
remaining list items.

When user scroll's over the view, if the view is not updated/
initialized with the new list item, it will render the old indexed
item(s), thus user will see the repeatation of old items.

This fix will update the view by initializing the item preference
from the adapter and returns the updated view based on new position.

CRs-fixed: 524897

Change-Id: I9ae25f0a711bc641fa405e85292db2c1af95e8a1