Gallery2: fix failed to play GIF image after screen rotated
The ViewGifImage activity will be recreated if user
rotate the phone.So the onStop method will be called.
Then onStop will call finish() method to destory this
activity.So gallery can't continue playing the GIF after
screen rotated.

This bug can be fixed by cutting the finish() from onStop.
If we do so,the gif will be replayed again.But the expected
result is that gallery should continue playing this gif.
So we add screenSize attribute into AndroidManifest to make
ViewGifImage won't be recreated after user rotate the phone.

CRs-Fixed: 543174

Change-Id: I373961ddfe3c95973de3ab146d5f3dcbc1781b8a