Camera: Synchronising video functionality of errant usecase.
When user switches to VideoModule and executes the capture, stops it
immediately, after several such attempts the media recorder enters
into illegal state, in such situation, if user still continues the
same behaviour, app will be flooded with the onShutter events which
will lead to ANR, and further similar trials will lead to crash.

Here, we are synchronizing the behaviour(user events)by checking the
entry/exit body of majorly hit functions so that app will not be
flooded with user events.

Change-Id: I3b7e1edad34607208cd597b37402e4a2192946ef
CRs-fixed: 534762