CameraApp: Updating the shutter button state based on active pie menu.
When user opens the pie renderer menu to switch the camera from
front to back or vice-verca, the shutter button will be inactive,
still sometimes the user tries to tap the button
forcefully/rigorously. In such situation the module tries to start
the mediarecorder without the camera switch operation getting
completed, which leads to ANR/ Crash.

With this fix, we are controlling the shutter click operation when
the switch option is still open.

Change-Id: I94c2d7113a90f56b6626a8d9322bd093d72754d7
CRs-fixed: 546725