CameraApp: Correction for float to int round off logic.
When preview size is small, the zoom circles calculation for
consecutive levels (58 and 59) was falling with in the range of
single digit. So float to int type casting was always falling down
to previou digit, hence user see 3.8 as the maximum zoom level in
front camera. This logic adds 0.5 before converting from float to
int.

Change-Id: I40f9ce14bba044a3e1802ee98a2d6669680970a2
CRs-fixed: 549781