Gallery2: fix cannot back to share in PhotoView for GIF image
For GIF type image,if user single tap on its thumbnail,will view
GIF image no matter the click position is or not in the picture
center. Then when action bar hides after timeout,user will never
back to share interface by single tap on the picture for it will
always come into playing this GIF.

We change to only play this GIF when the click position is in the
center of picture.

Change-Id: Id1ac38ebaa0193cec0f0fa83c73e67d4b8107c0e
CRs-Fixed: 551591