Gallery2: Fix save button disabled with Negative or AutoColor effect
The image filter name was set by resource id in FilterShowActivity,
it would be translated when the system language was not English.
But in the function hasModifications, we check the filter name by hard
code in English, such as "Negative" and "Autocolor".
That caused hasModifications function return false when we modified a
picture with these effect in the Chinese context.

Keep the filter name in English, only get the localized string for display.
Change "AutoColor" to "WBalance" in hasModifications, since its filter
name is "WBalance".

CRs-Fixed: 552014

Change-Id: I7d9858f85a0112b23fa19afff6cea41ac1aaaf68