Gallery2: seek to the closest sync frame instead of the seek position
When seeking MediaExtractor to a position, audio track is sought to
the exact position but video tract is sought to the nearby sync frame.

With this patch, while trimming videos, gallery first finds the sync
frame closest to the seek position, then seeks to the sync frame. So
audio track and video track are sought to the same position.

Change-Id: I390eb2064be8de62cb04bd25428c1a7354a4d138
CRs-Fixed: 552131