Gallery2: fix some gif picture display abnormal
If the disposal method of current frame is Restore to
Previous, it should request to restore to the state of
the previous frame whose disposal method is Unspecified
or Do Not Dispose, but it restore to previous frame
directly. This make the residual image phenomenon happen.

If the disposal method is Restore to Previous, we change to
restore the previous Unspecified or Do Not Dispose frame.

CRs-Fixed: 557343

Change-Id: I9ea572effa280610f745919606915f3786232bb7