Gallery2:Add judgement of sStepOptionArray in VideoSettingsActivity
In VideoSettingsActivity, after step option of 3/6 seconds is checked,
there is a possiblily of NullPointerException happen, for the array of
StepOption is not initialized successful.

Before the function of setSelectedStepOption is called, we need to
have a check of the Array of step option, if this array is null,
just dismiss the dialog.

CRs-fixed: 583086

Change-Id: If07cca4e9b4822300f423b2113a5c03a4f307bf1