Gallery2: fix screen flashing after rotating pictures
After rotating a picture, it needs to redraw the bitmap,
but the bitmap can not be available immediately, and it will
draw a rectange of placeholder color (gray) first, then it
redraw the bitmap, this make it seems like the screen flashing.

We draw the placeholder color only if the texture is null.

CRs-Fixed: 570056

Change-Id: I28bc65ba674906f39eae3d872988fe96511a3ecb