Gallery: Optimization getDrawable in ImageFilterBorder
When edit picture with both border and curves, Gallery will
force close for OutOfMemoryError if the curve is drap fast.
Because the border will be decode from resource whenever image
is render or compute, so decoder will be called too fast
and too many times during drapping.

I use a static Bitmap to save the decoded image of border,
if image is set a same border with last time, the static
Bitmap of border will be set directly instead of a decoding.
When request rendering, I catch the OOM exception.

CRs-Fixed: 573620

Change-Id: I502d06d945dfe8967e01af2467a958fd6f6c9e18