MoviePlayer: fix background issue when loading a video
The video's width and height may be set to 0 when it's prepared,
which makes the player treat it as an audio file, and set the music
icon as background.

Do not set the bottom panel if the player is not in playback state.

(cherry picked from commit a6facd923d07a66afa71e6cefa32ec8fc95bdb7b)

Change-Id: I7c2e3caf54b11b439419c22d9adb4e5eb205c706