Camera: Block HFR option for Video HDR,DIS,Timelapse
Video HDR can’t work with HFR and that caused AEC behaviour change.
Similarly DIS and Timelapse feature will not work with HFR.

Fix Made: User'll be notified that HFR/HSR selection is not allowed
for HDR,DIS,Timelapse mode. After that, HFR selection will be set to
off and the option will be disabled. It'll not be vice versa as
HDR,DIS and Timelapse has higher priority than HFR.

Change-Id: I003b3648ba87331db6595e37a10b3d9934f1a493
CRs-Fixed: 652078