Gallery2: Fix menu display inconsistent with its cluster
The operation order is not correct.We should first execute
the change of cluster,then display menu title according to
the cluster.But we do the opposite operation,which cause the
menu will always display last cluster‘s corresponding menu
and cause this bug happened.

Modify the operation order.

CRs-fixed: 619812

Change-Id: I6cd3c82e065cef5a463e06295bf332e1459500e3