Gallery3D: Handle mute failure for unsupported clips
If an exception is thrown due to failure of startMute(), a Toast
message is printed within the context of a worker thread, raising
an exception. Propagate the mute success status within the thread
handler to ensure that the Toast message can be correctly shown.

Also only trigger an intent to show the muted clip if mute succeeded.

Change-Id: Iae69e6b437ecad1f8936ecb36a2f3969dd15b549